Made v1,v2 double
Made the vertex index values a double integer. Better when working with
larger vertices.

+++ generator/src/main/java/ca/mcmaster/cas/se2aa4/a2/generator/CustomSegments.java
@@ -8,8 +8,8 @@ import java.awt.*;
 public class CustomSegments {
 
     private static final String THICKNESS = "3";
-    int v1;
-    int v2;
+    double v1;
+    double v2;
     Property colourProp;
     Property thicknessProp;
     Segment segment;
